Relocate `ParentBlockReferences` fore create dependency
The `service.rs` file is going to be updated to match the adjustments and improvements
that have been made in the `marketplace_builder_core`.  In order to prepare for this
appropriately `ParentBlockReferences` needs to be relocated os that it can also be
referenced within the `service.rs` file.
